+++ Example/Pods/CommonUtils/CommonUtils/Classes/Mapping/MappingError.swift
@@ -1,0 +1,13 @@
+//
+//  MappingError.swift
+//  Pods
+//
+//  Created by Ruben Pelzer on 20-06-17.
+//
+//
+
+import Foundation
+
+public enum MappingError: Error {
+    case jsonInvalid
+}